moving windower into own directory
