+++ basic demo/Subtitle System/Debug/Subtitle System.Build.CppClean.log
@@ -2,8 +2,8 @@ e:\temp\basic demo\subtitle system\debug\vc143.pdb
 e:\temp\basic demo\subtitle system\debug\vc143.idb
 e:\temp\basic demo\subtitle system\debug\window.obj
 e:\temp\basic demo\subtitle system\debug\main.obj
-e:\temp\basic demo\subtitle system\debug\subtitle system.ilk
 e:\temp\basic demo\debug\subtitle system.exe
+e:\temp\basic demo\subtitle system\debug\subtitle system.ilk
 e:\temp\basic demo\debug\subtitle system.pdb
 e:\temp\basic demo\subtitle system\debug\subtitle system.tlog\cl.command.1.tlog
 e:\temp\basic demo\subtitle system\debug\subtitle system.tlog\cl.read.1.tlog
