+++ basic demo/Subtitle System/windower/Window.cpp
